+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### [0.3.18] - 2023-02-08
+### Fixed
+- Fumble macro.
+
+### [0.3.17] - 2023-02-08 [Yanked]
+### Added
+- Fumble macro.
+
 ### [0.3.16] - 2023-01-22
 ### Changed
 - Skill macros respect smart select.